requirements.txt
streamlit==1.33.0
pandas==2.2.1
plotly==5.20.0
kafka-python==2.0.2
